write_frame_image() implemented
I was just trying to use python-theora to encode a series of images as a video,
i have them loaded in PIL Images. so first i needed to convert them to YCbCr,
and to an array. it would be nice if this would be part of python-theora
itself. so in addition to write_frame_array have write_frame_image which would
take a PIL image and add it to the video.